\name{nir}
\alias{nir}
\docType{data}
\title{Near-infrared data with temperature effects}
\description{A data object containing near-infrared spectra of ternary
  mixtures of ethanol, water and iso-propanol, measured at five
  different temperatures (30, 40, ..., 70 degrees Centigrade). } 
\references{F. Wulfert , W.Th. Kok, A.K. Smilde: Anal. Chem. 1998, 1761-1767}

\keyword{datasets}
